Deploy via CI to Cloud Run
This commit includes the configuration required to deploy the service to
Cloud Run.

== Design Notes
=== Container

While unnecessary as the container is currently being built by docker
hub, this commit also includes configuration to rebuild the container.

This will be useful as a kind of "proxy time" -- by the time this
container is built, the upstream container should also have been built.

Additionally, the goal is to move all containers into this pipeline and
not use the docker hub builder.

=== Cloud Run (via Plugin)

The author chose to deploy this service on Cloud Run as it is an easy to
start with, highly available and standards compatible deployment runtime
(i.e. it consumes a container).

It should implement the `knative` interface and use a fairly predictable
set of technologies.

Whether this is suitable for the long term viability of the service is
unclear.